添加xml库和优化zip库 1、添加xml库支持 2、zip功能从quanta剥离，独立成库 3、lmake功能扩展，支持多源文件目录，优化模板生成 4、修复部分编译警告 5、excel区域获取优化

+++ extend/lbson/src/bson.h
@@ -235,7 +235,7 @@ namespace lbson {
             if (depth > max_bson_depth) {
                 luaL_error(L, "Too depth while encoding bson");
             }
-            size_t raw_len = lua_rawlen(L, -1);
+            int raw_len = lua_rawlen(L, -1);
             bson_type type = check_doctype(L, raw_len);
             write_key(type, key, len);
             if (type == bson_type::BSON_DOCUMENT) {
